systems-requirements.txt
python-numpy
python-opencv
git
curl
libdc1394-22
libjpeg-turbo-progs
graphicsmagick
libgraphicsmagick++3
libgraphicsmagick-q16-3
libboost-python-dev
gifsicle
ffmpeg
